% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertest.R
\name{annotePathwayToFather}
\alias{annotePathwayToFather}
\title{Find Pathway Fathers}
\usage{
annotePathwayToFather(pathways, graphiteDB, hierarchy)
}
\arguments{
\item{pathways}{vector of pathway names}

\item{graphiteDB}{graphite DB object}

\item{hierarchy}{a graph object with the pathway hierarchy}
}
\value{
a vector of the fathers' names
}
\description{
Given the hierarchy of the pathways, this formula finds the fathers of the
respective pathway (e.g. pathway: "PI3K Cascade"; father:
"Signaling Pathways")
}
