% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{availableOmicMethods}
\alias{availableOmicMethods}
\title{Get available Omics Summarizing Methods}
\usage{
availableOmicMethods()
}
\value{
character vector with the implemented methods.
}
\description{
Gives a vector of the available methods to summarize omics.
}
