% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{availableOmicsMethods}
\alias{availableOmicsMethods}
\title{Get available Omics Summarizing Methods}
\usage{
availableOmicsMethods()
}
\value{
character vector with the implemented methods.
}
\description{
Gives a vector of the available methods to summarize omics.
}
