% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicsGlmTest.R
\name{checkFormula}
\alias{checkFormula}
\title{Check formulas}
\usage{
checkFormula(
  baseFormula,
  nullModel,
  classAnnot,
  autoCompleteFormula,
  omicsObj,
  paired,
  patientCol
)
}
\arguments{
\item{baseFormula}{model formula to be used for the test. It should be
written as 'classes ~ ', while 'classes' being the column name for the class
labels}

\item{nullModel}{the null model formula. It should be written the same as
the baseFormula, followed by ' 1'. (e.g. 'classes ~ 1')}

\item{classAnnot}{a \code{data.frame} with the class annotation. It is necessary
at least a column with the classes labels, and the row.names as the samples
labels}

\item{autoCompleteFormula}{a logical value, default TRUE.
If TRUE, it autocompletes the formula used to fit generalized linear
models function  using all the available covariates (omics)}

\item{omicsObj}{object of class \code{Omics}}

\item{paired}{whether samples are paired or not. Default FALSE. If TRUE,
a specific column should be present in \code{colData} specifying paired samples}

\item{patientCol}{name of the column to be searched in \code{colData} specifying
paired samples. Default "patient"}
}
\value{
NULL. Stop function if one of the formulas is not correct.
}
\description{
Check input baseFormula and nullModel formula are correct.
}
