% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertest.R
\name{computeOmicsIntersections}
\alias{computeOmicsIntersections}
\title{Compute Omics Intersections}
\usage{
computeOmicsIntersections(multiPathwayReportData, pvalueThr = 0.05,
  zscoreThr = 0.05, resampligThr = NULL, excludeColumns = NULL)
}
\arguments{
\item{multiPathwayReportData}{data.frame, the output of 
the \code{\link{multiPathwayReport}} or \code{\link{multiPathwayModuleReport}} functions.}

\item{pvalueThr}{numeric indicating the cut-off for overall pvalue}

\item{zscoreThr}{numeric indicating the cut-off for covariates coefficients}

\item{resampligThr}{numeric the number of success in the resamplig procedure}

\item{excludeColumns}{a vector of characters listing the columns of 
\code{multiPathwayReportData} object to be excluded by the analysis. 
In the case \code{multiPathwayReportData} derives from \code{\link{multiPathwayModuleReport}} 
you should set \code{excludeColumns = c("pathway","module")}.}
}
\value{
a list of pathway omics intersection
}
\description{
Compute Omics Intersections
}
