% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{convertPathway}
\alias{convertPathway}
\title{A generic function to convert pathway}
\usage{
convertPathway(graph, useTheseGenes)
}
\arguments{
\item{graph}{a graphNEL object}

\item{useTheseGenes}{list of genes to be used}
}
\value{
NULL. No value is returned
}
\description{
A generic function to convert pathway
}
