% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{createCoxObj}
\alias{createCoxObj}
\title{Create Cox Object}
\usage{
createCoxObj(colData, moView)
}
\arguments{
\item{colData}{colData from multiOmic object}

\item{moView}{modulesView or pathView from multiOmicsModules or
multiOmicsPathway object}
}
\value{
data.frame, samples in the rows, covariates in the columns
}
\description{
Create the coxObj from the covariates used in the test
}
