% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{createDataModule}
\alias{createDataModule}
\title{Create Data Module}
\usage{
createDataModule(omic, multiOmicObj)
}
\arguments{
\item{omic}{modulesView or pathView object}

\item{multiOmicObj}{object of class 'Omics'}
}
\value{
matrix, genes in the rows, samples in the columns
}
\description{
Extract sub-matrix for the genes of a module or pathway from data matrix of a specific omic
}
