% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModuleView.R
\name{createMOMView}
\alias{createMOMView}
\title{Create the list of covariates that are going to be tested}
\usage{
createMOMView(omicsObj, genes)
}
\arguments{
\item{omicsObj}{Omics object}

\item{genes}{genes to be used}
}
\value{
list with
1 reduced representation of the omics
2 sdev
3 loadings or eigenvector
4 usedGenes
5 method
6 namesCov
7 omicName
}
\description{
Create the list of covariates that are going to be tested
}
