% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyGraphModification.R
\name{createBiClasses}
\alias{createBiClasses}
\alias{createBinaryMatrix}
\alias{keepFirstOccurrence}
\title{From continous to discrete classes.}
\usage{
createBiClasses(coxObj, covs, doChecks = TRUE, labels = c("low",
  "high"), minprop = 0.1)

createBinaryMatrix(discrete, markAs1 = c("high", "TRUE"))

keepFirstOccurrence(m, whichCol)
}
\arguments{
\item{coxObj}{cox like object with days and status plus covariates}

\item{covs}{the covariates to use}

\item{doChecks}{if TRUE perform checks on min proportion}

\item{labels}{the labels of the two classes}

\item{minprop}{minimal proportion}

\item{discrete}{discrete version of the covariates (columns)}

\item{markAs1}{the discrete values associated to 1}

\item{m}{a matrix}

\item{whichCol}{the column to select the first occurrence}
}
\value{
discrete version of the selected 'covs'.

binary matrix

matrix without duplicates.
}
\description{
From a continous version of covariates (columns) to 2 discrete classes.

From a discrete version of covariates (columns), following markAs1 the function creates the binary version.

From a matrix or data frame keeps onòy the first occurrence according to a column.
}
\examples{
  dummy <- matrix(c("high","high","low","TRUE","low","low","high","FALSE"), nrow=4)
  createBinaryMatrix(dummy)
  dummy <- matrix(c("a","b","a","c",1,2,3,4), nrow=4)
  colnames(dummy) <- c("gene", "value")
  keepFirstOccurrence(dummy,1)
  keepFirstOccurrence(dummy,"gene")
  
}
