% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy-datasets-creator.R
\name{dummy_methylation_like_dataset}
\alias{dummy_methylation_like_dataset}
\alias{dummy_methylation_like_flat_dataset}
\alias{create_met_cluster_dict}
\alias{dummy_expression_like_dataset}
\alias{dummy_cnv_like_dataset}
\title{Function to create some fake dataset for testing purpose}
\usage{
dummy_methylation_like_dataset(seed = 1234)

dummy_methylation_like_flat_dataset(seed = 1234)

create_met_cluster_dict(dataset)

dummy_expression_like_dataset(seed = 1234)

dummy_cnv_like_dataset(seed = 1234)
}
\arguments{
\item{seed}{the seed number for reproducibility of random numbers}

\item{dataset}{a dummy dataset for the dummy dict}
}
\value{
dataset

dict
}
\description{
Create a dataset with 200 patients and 5 genes:
  3 genes contains 150/50 ratio difference with mean obtained with 
    runif (n1 , 0, 0.2) / runif(n2, 0.7, 0.99)
  2 genes 75/125 ratio difference in the opposite direction obtained with
    runif (n3 , 0.6, 1) / runif(n4, 0.2, 0.4)

Create a dataset with 200 patients and 1 genes:
  generated by 191 0.3 and 9 0.4

Create a dataset with 200 patients and 5 genes:
  3 genes contains 150/50 ratio difference with mean obtained with 
    rnorm (n1 , 5, 1) / rnorm(n2, 8, 1)
  2 genes 75/125 ratio difference in the opposite direction obtained with
    rnorm (n3 , 9, 1) / rnorm(n4, 2, 0.4)

Create a dataset with 200 patients and 5 genes:
  3 genes contains 150/50 ratio difference with values obtained by
    sampling c(-1,0,0,0,1,1) and c(-2,0,0,0,0,2,2) for low/high
  2 genes 75/125 ratio difference in the opposite direction obtained by
    sampling c(-2,0,0,0,0,0,2) and c(-1,0,0,0,0,0,1) for high/low
}
