% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingFunction.R
\name{selectStablePathwaysModules}
\alias{selectStablePathwaysModules}
\alias{getPathwaysModulesSuccess}
\alias{addResamplingCounts}
\title{Select stable pathway modules}
\usage{
selectStablePathwaysModules(perms, moduleSummary, success = 90,
  col = "pvalue")

getPathwaysModulesSuccess(perms, moduleSummary, col = "pvalue",
  thr = 0.05)

addResamplingCounts(moduleSummary, resamplingCounts)
}
\arguments{
\item{perms}{a list of perm objects}

\item{moduleSummary}{summary of the modules}

\item{success}{number of success}

\item{col}{the name of the column}

\item{thr}{the threshold for significance}

\item{resamplingCounts}{the counts of success}
}
\value{
the subset of stable modules

the counts of success

a module summary with reampling counts
}
\description{
Select stable pathway modules

Count the resampling success

Add resampling counts to module summary
}
