% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis-functions.R
\name{extractBadPrognosisProfile}
\alias{extractBadPrognosisProfile}
\title{Extract the worst profile}
\usage{
extractBadPrognosisProfile(coxDiscrete)
}
\arguments{
\item{coxDiscrete}{the discrete table genes x samples}
}
\value{
a data frame with covatiate (cov) and its profile associated to the bad prognosis
}
\description{
Extract the profile with the lowest survival time.
}
