% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRiskTable.R
\name{extractRiskTable}
\alias{extractRiskTable}
\title{Extrac the risk table from survminer object}
\usage{
extractRiskTable(survminerObj)
}
\arguments{
\item{survminerObj}{a survminer object}
}
\description{
Extrac the risk table from survminer object
}
