% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSummary.R
\name{extractSummaryFromNumberOfEvents}
\alias{extractSummaryFromNumberOfEvents}
\title{Extract Summary Binary from MultiOmics Objects}
\usage{
extractSummaryFromNumberOfEvents(omic, moduleCox, n = 3, minprop = 0.1)
}
\arguments{
\item{omic}{a summarized omic}

\item{moduleCox}{the coxObj of the interesting module}

\item{n}{maximum number of features to retrieve}

\item{minprop}{the minimal proportion of cutp}
}
\value{
Meant for internal use only. The summary for omic summarized using clusters
\item{sigModule}{the original data for significant features}
\item{discrete}{the discrete version of the significant covariates converted (when needed)
  into the discrete version}
\item{subset}{data.frame(row.names=names(topGenes), metClust=topGenes)}
\item{covsConsidered}{the name of the considered omic}
}
\description{
Given an omic summarized by "summarizeToBinaryEvents" extract the most important features.
}
