% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMGlmTest.R
\name{glmTest}
\alias{glmTest}
\title{Two-classes glm test.}
\usage{
glmTest(data, fullModelFormula, nullModelFormula)
}
\arguments{
\item{data}{data}

\item{fullModelFormula}{complete model}

\item{nullModelFormula}{null model formula}
}
\value{
Two-classes glm test results
}
\description{
Two-classes glm test.
}
