% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessInvolvement.R
\name{guessInvolvement}
\alias{guessInvolvement}
\title{Guess the most influent features from MultiOmics Survival Results.}
\usage{
guessInvolvement(pathway, moduleNumber, loadThr = 0.6, n = 3, atleast = 1)
}
\arguments{
\item{pathway}{MultiOmicsModule from a pathway.}

\item{moduleNumber}{the module number}

\item{loadThr=0.6}{the leading threshold to select genes (PCA only)}

\item{n=3}{the maximum number of genes to retrive (cluster and binary only)}

\item{atleast=1}{the minimum number of features to select (PCA only)}
}
\value{
For each omic analyzed a list that is the summary for omic summarized using the setted method: pvalues are present only for cluster method.
\item{sigModule}{the original data for significant features}
\item{discrete}{the discrete version of the significant covariates converted (when needed) into the discrete version}
\item{subset}{data.frame(row.names=names(topGenes), metClust=topGenes)}
\item{pvalues}{Kruskal Wallis pvalues of the selected features}
\item{covsConsidered}{the name of the considered omic}
}
\description{
Given a pathway analyzed by MultiOmicsModuleSurvivalTest it retrieve for each omic the most influent fetures.
}
