% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessInvolvement.R
\name{guessInvolvementPathway}
\alias{guessInvolvementPathway}
\title{Guess the most influent features from MultiOmics Survival Results.}
\usage{
guessInvolvementPathway(pathway, loadThr = 0.6, n = 3, atleast = 1,
  min_prop_pca = 0.1, min_prop_events = 0.1)
}
\arguments{
\item{pathway}{MultiOmicsPathway from a pathway.}

\item{loadThr}{the leading threshold to select genes (PCA only)}

\item{n}{the maximum number of genes to retrive (cluster and binary only)}

\item{atleast}{the minimum number of features to select (PCA only)}

\item{min_prop_pca}{the minimal proportion to compute the pca classes}

\item{min_prop_events}{the minimal proportion to compute the event classes}
}
\value{
For each omic analyzed a list that is the summary for omic summarized using the setted method: pvalues are present only for cluster method.
\item{sigModule}{the original data for significant features}
\item{discrete}{the discrete version of the significant covariates converted (when needed) into the discrete version}
\item{subset}{data.frame(row.names=names(topGenes), metClust=topGenes)}
\item{pvalues}{Kruskal Wallis pvalues of the selected features}
\item{covsConsidered}{the name of the considered omic}
}
\description{
Given a pathway analyzed by MultiOmicsPathwaySurvivalTest it retrieve for each omic the most influent fetures.
}
