% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertest.R
\name{match_pathway_to_fathers}
\alias{match_pathway_to_fathers}
\title{Remove pathways with no id in graphite}
\usage{
match_pathway_to_fathers(omicClass, omicsClasses2pathways,
  omicsClasses2fathers)
}
\arguments{
\item{omicClass}{the name of an omic class (or omic combination)}

\item{omicsClasses2pathways}{vector of pathway names}

\item{omicsClasses2fathers}{vector of pathway names}
}
\value{
data frame
}
\description{
This function id risky to use. It performs pairing without
checking proper pairs.
}
