% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis-functions.R
\name{multiOmicsModuleInterAnalysis}
\alias{multiOmicsModuleInterAnalysis}
\title{Multi Omics Module Inter Module Analysis}
\usage{
multiOmicsModuleInterAnalysis(moduleSummary, momTestObj, zscoreThr = 0.05)
}
\arguments{
\item{moduleSummary}{the table that summarized the module results}

\item{momTestObj}{the list of MOM results}

\item{zscoreThr}{significance threshold}
}
\value{
a list
\item{sigOmicsPart}{for each Omic the significant}
\item{pvaluesSummary}{for each omic, the minimum pvalue}
\item{allGenes}{the genes used}
}
\description{
Given a list of pathway results along with the summary table, it summaizes all module results
}
