% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmic-classes.R
\docType{class}
\name{MultiOmicsPathway-class}
\alias{MultiOmicsPathway-class}
\alias{MultiOmicsPathway}
\alias{MultiOmicsPathway-class}
\alias{plotMultiOmicsPathway,MultiOmicsPathway-method}
\title{Multi Omics Pathway that extends \code{"list"} class.}
\usage{
plotMultiOmicsPathway(object)

\S4method{plotMultiOmicsPathway}{MultiOmicsPathway}(object)
}
\description{
This class organize the results of the Multi Omics Module Survival Test analysis.

Method plotMultiOmicsPathway.
}
\section{Slots}{

\describe{
\item{\code{pvalue}}{numeric, the pvalues of the whole module.}

\item{\code{zlist}}{a vector of all the zs of the covariates.}

\item{\code{coxObj}}{a data.frame used for the coxph model.}

\item{\code{pathView}}{a list, for each omic the pathway data, the method used and the covariate analyzed.}

\item{\code{formulas}}{a list, for each module the character of the formula used in the coxph.}

\item{\code{pathData}}{a list, for each omics the data analyzed. Redundant.}

\item{\code{graphNEL}}{the graphNEL version of the pathway used in the analysis.}

\item{\code{title}}{the name of the pathway.}
}}

