% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicsSurvivalTests.R
\name{multiOmicsSurvivalModuleTest}
\alias{multiOmicsSurvivalModuleTest}
\title{Compute Multi Omics Survival in Pathway Modules}
\usage{
multiOmicsSurvivalModuleTest(omicsObj, graph, annot,
  survFormula = "Surv(days, status) ~", autoCompleteFormula = T,
  useThisGenes = NULL, pathName = NULL, robust = FALSE,
  include_from_annot = F)
}
\arguments{
\item{omicsObj}{Object of class 'Omics'}

\item{graph}{a pathway in graphNEL, Pathway or geneset format.}

\item{annot}{survival annotation: days and status (0,1). Row.names are samples. Additional covariates can be passed as columns (use include_from_annot).}

\item{survFormula}{Formula to compute survival}

\item{autoCompleteFormula}{logical. If TRUE autocomplete the survFormula using all the available covariates}

\item{useThisGenes}{vector of genes used to filter pathways}

\item{pathName}{title of the pathway. If NULL and graph is "Pathway" graph@title is used as title}

\item{robust}{should be used the robust mode for cox}

\item{include_from_annot}{compute cox using additional covariates from annot}
}
\value{
MultiOmicsModules object
}
\description{
Performs survival analysis using an 'Omics' object. The pathway (graph) used is decomposed in modules (cliques) using graph theory.
}
