% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicsSurvivalTests.R
\name{multiOmicsSurvivalPathwayTest}
\alias{multiOmicsSurvivalPathwayTest}
\title{Compute Multi Omics Survival in Pathways}
\usage{
multiOmicsSurvivalPathwayTest(omicsObj, graph, daysStatus,
  survFormula = "Surv(days, status) ~", autoCompleteFormula = T,
  useThisGenes = NULL, pathName = NULL, robust = FALSE)
}
\arguments{
\item{omicsObj}{Object of class 'Omics'}

\item{graph}{a pathway in graphNEL, Pathway or geneset format.}

\item{daysStatus}{survival annotation: days and status (0,1). Row.names are samples}

\item{survFormula}{Formula to compute survival}

\item{autoCompleteFormula}{logical. If TRUE autocomplete the survFormula using all the available covariates}

\item{useThisGenes}{vector of genes used to filter pathways}

\item{pathName}{title of the pathway. If NULL and graph is "Pathway" graph@title is used as title}

\item{robust}{should be used the robust mode for cox}
}
\value{
MultiOmicsPathway object
}
\description{
Performs topological survival analysis using an 'Omics' object.
}
