% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiOmicsTopo}
\alias{multiOmicsTopo}
\title{Omics class object with TCGA ovarian data for topological analysis}
\format{
\subsection{\code{multiOmicsTopo}}{

An Omics with 4 omics:
\describe{
\item{exp}{Matrix with 151 rows and 50 columns of RNA expression values}
\item{met}{A matrix with 178 rows and 50 columns of methylation data with
probes clustered}
\item{mut}{A matrix with 107 rows and 50 columns of mutation counts}
\item{cnv}{A matrix with matrix with 145 rows and 50 columns of copy
number}
...
}
}
}
\usage{
data('multiOmicsTopo')
}
\description{
An \code{Omics} class object containing data from TCGA ovarian cancer.
The data are the same as in \code{\link{multiOmics}} object.
Arguments in \code{specificArgs} slot have been set to efficiently run a
topological pathway analysis, i.e., the topological method is used for PCA
and shrink parameter is set to TRUE.
This method can't be used for analyses on modules.
}
\keyword{datasets}
