% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiOmicsTwoClass}
\alias{multiOmicsTwoClass}
\title{Omics class object with TCGA ovarian data for two-class analysis}
\format{
\subsection{\code{multiOmicsTwoClass}}{

An Omics object with 4 omics:
\describe{
\item{exp}{Matrix with 101 rows and 50 columns of RNA expression values}
\item{met}{A matrix with 97 rows and 50 columnsof methylation data with
probes clustered}
\item{mut}{A matrix with 55 rows and 50 columns of mutation counts}
\item{cnv}{A matrix with matrix with 101 rows and 50 columns of copy
number}
...
}
}
}
\usage{
multiOmicsTwoClass
}
\description{
A Omics class object containing data from TCGA ovarian cancer. The TCGA data
was manually selected and preprocessed. It contains 4 omics: expression,
methylation, mutation, and copy number variation. Additionally, it contains
specific arguments to perform the dimensionality reduction and colData.
}
\keyword{datasets}
