% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicsGlmTest.R
\name{multiOmicsTwoClassesModuleTest}
\alias{multiOmicsTwoClassesModuleTest}
\title{Compute Multi Omics two-class in Pathway Modules}
\usage{
multiOmicsTwoClassesModuleTest(
  omicsObj,
  graph,
  classAnnot,
  baseFormula = "classes ~ ",
  autoCompleteFormula = TRUE,
  useThisGenes = NULL,
  nullModel = "classes ~ 1",
  pathName = NULL
)
}
\arguments{
\item{omicsObj}{Object of class 'Omics'}

\item{graph}{a pathway in graphNEL, Pathway or geneset format.}

\item{classAnnot}{class annotation: name of the classes, row.names are samples}

\item{baseFormula}{model to be used for the test}

\item{autoCompleteFormula}{logical. If TRUE autocomplete the survFormula using all the available covariates}

\item{useThisGenes}{vector of genes used to filter pathways}

\item{nullModel}{the null model}

\item{pathName}{title of the pathway. If NULL and graph is "Pathway" graph@title is used as title}
}
\value{
MultiOmicsModule object
}
\description{
Performs topological two-class analysis using an 'Omics' object on pathway module.
}
