% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.R
\name{Omics}
\alias{Omics}
\title{#'
setGeneric("initialize", function(.Object,experime, modelInfo, specificArgs,pathResult) standardGeneric("initialize"))}
\usage{
Omics(experime, modelInfo, specificArgs, pathResult)
}
\description{
#'
setMethod("initialize",
signature(.Object = "Omics"),\preformatted{      function(.Object,experime, modelInfo, specificArgs,pathResult) \{
         
        .Object <- callNextMethod()
        .Object
      \}
}

)
}
