% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarianDataset}
\alias{ovarianDataset}
\title{ExperimentList class object with TCGA ovarian data}
\format{
\subsection{\code{ExperimentList}}{

An ExperimentList with 4 omics:
\describe{
\item{exp}{Matrix with 101 rows and 50 columns of RNA expression values}
\item{met}{A matrix with 97 rows and 50 columnsof methylation data with
probes clustered}
\item{mut}{A matrix with 55 rows and 50 columns of mutation counts}
\item{cnv}{A matrix with matrix with 101 rows and 50 columns of copy
number}
...
}
}
}
\usage{
data("ovarianDataset")
}
\description{
An ExperimentList class object containing data from TCGA ovarian cancer.
The TCGA data was manually selected and preprocessed. It contains 4 omics:
expression, methylation, mutation, and copy number variation.
}
\keyword{datasets}
