% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyGraphModification.R
\name{plotGraphiteInCy}
\alias{plotGraphiteInCy}
\title{Plot graph in Cytoscape}
\usage{
plotGraphiteInCy(nattributes, attributeClass, graph)
}
\arguments{
\item{nattributes}{a dataframe with the attribute. Colnames equal to type}

\item{attributeClass}{a dataframe with class and class type}

\item{graph}{a graph}
}
\value{
invisble list
\item{gNEL}{the graphNEL object}
\item{cy}{the cytoscape connection}
}
\description{
Given the pathway and nodesAttributes create a graph in Cytoscape
}
