% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotMultiPathwayReport}
\alias{plotMultiPathwayReport}
\title{Summarize and plot pathways' info from a list of MultiOmicsPathway (MOP)}
\usage{
plotMultiPathwayReport(multiPathwayList)
}
\arguments{
\item{multiPathwayList}{MultiOmicsPathway list pathway object}
}
\description{
Given the list of MOPs, it plots the table.
}
