% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotPathwayHeat}
\alias{plotPathwayHeat}
\title{Plot heatmaps of the pathway by omics}
\usage{
plotPathwayHeat(pathway, sortBy = NULL, fileName = NULL,
  paletteNames = NULL, additionalAnnotations = NULL,
  additionalPaletteNames = NULL, withSampleNames = TRUE,
  fontsize_row = 10, fontsize_col = 1, nrowsHeatmaps = 3, h = 9,
  w = 7)
}
\arguments{
\item{pathway}{MultiOmicsPathway pathway object}

\item{sortBy}{a covariate to sort by}

\item{fileName}{optional filenames to save the plot}

\item{paletteNames}{three palettes}

\item{additionalAnnotations}{optional additional sample annotations}

\item{additionalPaletteNames}{optional additional colors for annotations}

\item{withSampleNames}{create also the samples names}

\item{fontsize_row}{size of the fonts for rows}

\item{fontsize_col}{like fontsize_row but for columns}

\item{nrowsHeatmaps}{magnification respect to annotation of sample (annotations take 1 row)}

\item{h}{the height of the plot}

\item{w}{the width of the plot}
}
\description{
Given the pathway, it creates the heatmaps of the mostly involved genes for each omic.
}
