% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactSmall}
\alias{reactSmall}
\title{PathwayList of pathways from Reactome}
\format{
\subsection{\code{reactSmall}}{

A PathwayList with Reactome pathways for hsapiens
\describe{
\item{entries}{Three reactome pathways with their nodes}
}
}
}
\usage{
reactSmall
}
\description{
A PathwayList with three pathways necessary for the analysis:
"Intrinsic Pathway for Apoptosis", "Opioid Signalling", and
"Mitochondrial protein import"
}
\keyword{datasets}
