% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseClip-graph-processing-functions.R
\name{removeSelfLoops}
\alias{removeSelfLoops}
\title{Remove self loops from a graphNEL}
\usage{
removeSelfLoops(graph)
}
\arguments{
\item{graph}{a graphNEL object}
}
\value{
a graphNEL object

#' @rdname  graph-processing
}
\description{
Remove the self loops that a present in the graph graphNEL object
}
