% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingFunction.R
\name{checkOrder}
\alias{checkOrder}
\alias{resolveAndOrder}
\alias{mergeCol}
\alias{filterExpr}
\alias{filterMultiOmicsForSamples}
\alias{resampling}
\title{Check if all the list object have the same order of pathway module}
\usage{
checkOrder(li)

resolveAndOrder(li)

mergeCol(li, col = "PC1", resolve = FALSE)

filterExpr(exp, samples)

filterMultiOmicsForSamples(MO, samples)

resampling(fullMultiOmics, survAnnot, pathdb, nperm = 100,
  pathwaySubset = NULL, nPatients = 3)
}
\arguments{
\item{li}{a list of summaries}

\item{col}{the column to merge}

\item{resolve}{weather to resolve the issues}

\item{exp}{a matrix}

\item{samples}{the vector of samples to select}

\item{MO}{a multiOmic object}

\item{fullMultiOmics}{a multiOmic object}

\item{survAnnot}{survival annotation}

\item{pathdb}{pathway dayabase}

\item{nperm}{number of permutation}

\item{pathwaySubset}{a list of pathways to resample}
}
\value{
a matrix

a filtered matrix

a filtered MultiOmics objects

list of the resampling tables of results
}
\description{
For internal use only

For internal use only

For internal use only

For internal use only

For internal use only

Resampling function
}
