% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingFunction.R
\name{resamplingPathway}
\alias{resamplingPathway}
\title{Resampling function for pathways}
\usage{
resamplingPathway(fullMultiOmics, survAnnot, pathdb, nperm = 100,
  pathwaySubset = NULL, nPatients = 3)
}
\arguments{
\item{fullMultiOmics}{a multiOmic object}

\item{survAnnot}{survival annotation}

\item{pathdb}{pathway dayabase}

\item{nperm}{number of permutation}

\item{pathwaySubset}{a list of pathways to resample}

\item{nPatients}{number of patients to remove for resampling}
}
\value{
list of the resampling tables of results
}
\description{
Resampling function for pathways
}
