% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingFunctions.R
\name{resamplingSurvival}
\alias{resamplingSurvival}
\title{Resampling function for survival analysis on modules}
\usage{
resamplingSurvival(
  fullMultiOmics,
  pathdb,
  nperm = 100,
  pathwaySubset = NULL,
  nPatients = 3
)
}
\arguments{
\item{fullMultiOmics}{a multiOmic object}

\item{pathdb}{pathway database}

\item{nperm}{number of permutations}

\item{pathwaySubset}{a list of pathways to resample}

\item{nPatients}{number of patients to remove for resampling}
}
\value{
list of the resampling tables of results
}
\description{
Resampling function for survival analysis on modules
}
