% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{showModule}
\alias{showModule}
\title{A generic function showing pathway's module info}
\usage{
showModule(object)
}
\arguments{
\item{object}{an object of class \code{MultiOmicsModules}}
}
\value{
NULL. No value is returned
}
\description{
A generic function showing pathway's module info
}
