% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseClip-pca-functions.R
\name{sparseCompPCs}
\alias{sparseCompPCs}
\title{Sparse PCA}
\usage{
sparseCompPCs(exp, shrink, k)
}
\arguments{
\item{exp}{a matrix}

\item{shrink}{logical, whether to shrink or not.}

\item{k}{the number of components to use}
}
\value{
a list with the following elements:
\item{x}{the computed PCs}
\item{sdev}{the standard deviation captured by the PCs}
\item{loadings}{the loadings}
}
\description{
Sparse PCA
}
