% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis-functions.R
\name{stripType}
\alias{stripType}
\alias{stripOmicsName}
\alias{addOmicsName}
\title{Strip feature type.}
\usage{
stripType(nms, type)

stripOmicsName(nms, name, sep = ".")

addOmicsName(nms, name, sep = ".")
}
\arguments{
\item{nms}{character vector of names}

\item{type}{the string type to be stripped}

\item{name}{the omic name}

\item{sep}{the separator}
}
\value{
a character vector of names stripped

a character vector of names stripped

a character vector of names stripped
}
\description{
Remove the type from the gene name.

Remove the omic prefix from the gene name.

Add the omic prefix to the gene name.
}
\details{
The type is marked with the separator ':' (\code{graphite} style)

The type is marked with the separator sep

The type is marked with the separator sep.
}
