% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis-functions.R
\name{summarizeColumnsByMask}
\alias{summarizeColumnsByMask}
\title{Multi Omics Module Inter Module Analysis}
\usage{
summarizeColumnsByMask(table, mask, thr = 0.05)
}
\arguments{
\item{table}{the table that summarized the module results}

\item{mask}{a factor to summarize. 'rm' marked columns are dropped}

\item{thr}{significance threshold}
}
\value{
a list
\item{table}{the original table without the modules without at least one covariate significant}
\item{sigMask}{for each module and omics, TRUE/FALSE for significance}
\item{sumPvalues}{pvalues summaries for each omics}
}
\description{
Given the results summary, it extract the lowest pvalue according to the sMask factor.
}
