% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{summarizeInClusterWithoutDictionary}
\alias{summarizeInClusterWithoutDictionary}
\title{Summarize Using Cluster Analysis with no dictionary to translate the matrix ids}
\usage{
summarizeInClusterWithoutDictionary(data, features, name = "clu",
  cliques = NULL)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{cliques}{the features organized in cliques. Only use for topology.}
}
\description{
Given a quantitative matrix it clusters the samples
}
