% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{summarizeToBinaryDirectionalEvents}
\alias{summarizeToBinaryDirectionalEvents}
\title{Summarize To Binary Directional Events}
\usage{
summarizeToBinaryDirectionalEvents(data, features, name = "dirBin",
  binaryClassMin = 10, eventThr = 2, cliques = NULL)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{binaryClassMin}{the minimum number of event to include the covariate}

\item{eventThr}{the absolute value to threshold an event}

\item{cliques}{the features organized in cliques. Only use for topology.}
}
\description{
Given a matrix it summarize the positive and negative in two vectors
}
