% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{summarizeToBinaryEvents}
\alias{summarizeToBinaryEvents}
\title{Summarize To Binary Events}
\usage{
summarizeToBinaryEvents(data, features, name = "cov", binaryClassMin = 10,
  cliques = NULL)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{binaryClassMin}{the minimum number of event to include the covariate}

\item{cliques}{the features organized in cliques. Only use for topology.}
}
\description{
Given a matrix it summarize to a 0 or 1
}
