% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{summarizeToNumberOfEvents}
\alias{summarizeToNumberOfEvents}
\title{Summarize To Binary Events}
\usage{
summarizeToNumberOfEvents(data, features, name = "event",
  cliques = NULL)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{cliques}{the features organized in cliques. Only use for topology.}
}
\description{
Given a matrix it summarize to a 0 or 1
}
