% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{summarizeWithPca}
\alias{summarizeWithPca}
\title{Summarize Using PCA}
\usage{
summarizeWithPca(data, features, name = "exprs", shrink = FALSE,
  method = "regular", cliques = NULL, maxPCs = 3, loadThr = 0.6)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{shrink}{shirnk or not the covariance matrix.}

\item{method}{either "regular", "sparse" or "topological"}

\item{cliques}{the features organized in cliques. Only use for topology.}

\item{maxPCs}{maximum number of pcs to consider}

\item{loadThr}{loading threshold}
}
\description{
Given a matrix it summarize to a 0 or 1
}
