% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseClip-pca-functions.R
\name{topoCompPCs}
\alias{topoCompPCs}
\title{Topological PCA}
\usage{
topoCompPCs(exp, shrink, cliques, k)
}
\arguments{
\item{exp}{a matrix}

\item{shrink}{logical, whether to shrink or not.}

\item{cliques}{the pathway topology summarized in a list of cliques}

\item{k}{the number of components to use}
}
\value{
a list with the following elements:
\item{x}{the computed PCs}
\item{sdev}{the standard deviation captured by the PCs}
\item{loadings}{the loadings}
}
\description{
Topological PCA
}
