% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertest.R
\name{annotePathwayToFather}
\alias{annotePathwayToFather}
\title{Convert pathways into a fathers}
\usage{
annotePathwayToFather(pathways, graphiteDB, hierarchy)
}
\arguments{
\item{pathways}{vector of pathway names}

\item{graphiteDB}{graphite DB object}

\item{hierarchy}{a graph object with the pathway hierarchy}
}
\value{
a vector of fathers names
}
\description{
Convert pathways into a fathers
}
