% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMSurvTest.R
\name{createMOMView}
\alias{createMOMView}
\title{Create the list of covariates that are going to be tested}
\usage{
createMOMView(omicsObj, genes)
}
\value{
list with
1 reduced representation of the omics
2 sdev
3 loadings or eigenvector
4 data module ## Consider adding here only genes)
5 Method
6 namesCov
7 OmicName
}
\description{
Create the list of covariates that are going to be tested
}
