% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{MultiOmicsModules-class}
\alias{MultiOmicsModules-class}
\alias{showModule,MultiOmicsModules-method}
\title{Multi Omics Modules.}
\usage{
\S4method{showModule}{MultiOmicsModules}(object)
}
\arguments{
\item{object}{an object of class MultiOmicsModules}
}
\description{
This class organize the results of the Multi Omics Module Survival Test analysis.
}
\section{Methods (by generic)}{
\itemize{
\item \code{showModule(MultiOmicsModules)}: shows module info

}}
\section{Slots}{

\describe{
\item{\code{alphas}}{numeric vector of the pvalues of all the modules.}

\item{\code{zlists}}{a list of all the zs of the covariates.}

\item{\code{coxObjs}}{a list with all the data.frame used for coxph of each module.}

\item{\code{modulesView}}{a list of module information: for each omic the module data, the method used and the covariate analyzed.}

\item{\code{modules}}{a list woth the genes that belong to the module.}

\item{\code{formulas}}{a list, for each module the character of the formula used in the coxph.}

\item{\code{title}}{the name of the pathway.}

\item{\code{analysis}}{The type of analysis done: survival or two-class.}
}}

