% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{MultiOmicsModules-class}
\alias{MultiOmicsModules-class}
\alias{showModule,MultiOmicsModules-method}
\title{Multi Omics Modules.}
\usage{
\S4method{showModule}{MultiOmicsModules}(object)
}
\arguments{
\item{object}{an object of class \code{MultiOmicsModules}}
}
\description{
<<<<<<< HEAD
This class organizes the results of the Multi-Omics Module Survival Test
analysis.
=======
This class organize the results of the Multi Omics Module Test analysis, in
which corresponds to one pathway decomposed into modules.
>>>>>>> origin/anna_t
}
\section{Methods (by generic)}{
\itemize{
\item \code{showModule(MultiOmicsModules)}: shows module info

}}
\section{Slots}{

\describe{
\item{\code{alphas}}{a numeric vector of the pvalues of all the modules.}

\item{\code{zlists}}{a list of numeric vectors with the zs of the covariates
for each module.}

\item{\code{modulesView}}{a list of module information: for each omic,
the name of the omic, the genes used, the method, the name of the
covariates analyzed and other specific information based on the omic.}

\item{\code{modules}}{a list with the genes that belong to the module.}

\item{\code{title}}{the name of the pathway.}

\item{\code{analysis}}{the type of analysis done: survival or two-class.}
}}

