% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{MultiOmicsPathway-class}
\alias{MultiOmicsPathway-class}
\alias{showPathway,MultiOmicsPathway-method}
\title{Multi Omics Pathway.}
\usage{
\S4method{showPathway}{MultiOmicsPathway}(object)
}
\arguments{
\item{object}{an object of class MultiOmicsPathway}
}
\description{
This class organize the results of the Multi Omics Module Survival Test analysis.
}
\section{Methods (by generic)}{
\itemize{
\item \code{showPathway(MultiOmicsPathway)}: shows module info

}}
\section{Slots}{

\describe{
\item{\code{pvalue}}{numeric, the pvalues of the whole module.}

\item{\code{zlist}}{a vector of all the zs of the covariates.}

\item{\code{coxObj}}{a data.frame used for the coxph model.}

\item{\code{pathView}}{a list, for each omic the pathway data, the method used and the covariate analyzed.}

\item{\code{title}}{the name of the pathway.}

\item{\code{analysis}}{The type of analysis done: survival or two-class.}
}}

