% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{MultiOmicsPathway-class}
\alias{MultiOmicsPathway-class}
\alias{showPathway,MultiOmicsPathway-method}
\title{Multi Omics Pathway.}
\usage{
\S4method{showPathway}{MultiOmicsPathway}(object)
}
\arguments{
\item{object}{an object of class \code{MultiOmicsPathway}}
}
\description{
This class organize the results of the Multi-Omics Pathway Survival Test
analysis.
}
\section{Methods (by generic)}{
\itemize{
\item \code{showPathway(MultiOmicsPathway)}: shows module info

}}
\section{Slots}{

\describe{
\item{\code{pvalue}}{a numeric vector of the pvalues of the pathways.}

\item{\code{zlist}}{a numeric vector with the zs of all the covariates.}

\item{\code{pathView}}{a list of pathway information: for each omic,
the name of the omic, the genes used, the method, the name of the
covariates analyzed and other specific information based on the omic.}

\item{\code{title}}{the name of the pathway.}

\item{\code{analysis}}{the type of analysis done: survival or two-class.}
}}

