% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTabulaReport.R
\name{multiPathwayModuleReport}
\alias{multiPathwayModuleReport}
\title{Summarize modules' info from a list of MultiOmicsModules objects (MOP)}
\usage{
multiPathwayModuleReport(multiPathwayModuleList, priority_to = NULL)
}
\arguments{
\item{multiPathwayModuleList}{a list of \code{MultiOmicsModules} objects
resulting from a multi-omics module test.}

\item{priority_to}{a vector with the covariates (omic name) that should go
first.}
}
\value{
a data.frame, modules in rows, overall pvalue of the coxph,
followed by covariates pvalues, in columns.
}
\description{
Given the list of MOMs, it creates the table.
}
