% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTabularReports.R
\name{multiPathwayModuleReport}
\alias{multiPathwayModuleReport}
\title{Provide a data.frame of pathways module test results from list of Multi Omics Module (MOM) objects}
\usage{
multiPathwayModuleReport(multiPathwayModuleList)
}
\arguments{
\item{multiPathwayModuleList}{a list of Multi Omics Modules resulting from a multi-omics module test.}
}
\value{
a data.frame, modules in rows, overall and covariate pvalues of the test in columns.
#'
}
\description{
Given the list of MOMs, it creates the table.
}
