% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTabularReports.R
\name{multiPathwayModuleReport}
\alias{multiPathwayModuleReport}
\title{Summarize pathways' Module info from a list of MultiOmicsModule (MOM) objects}
\usage{
multiPathwayModuleReport(multiPathwayModuleList)
}
\arguments{
\item{multiPathwayModuleList}{MultiOmicsModule list of pathway}
}
\value{
a data.frame with overal pvalue of the coxph, followed by covariates zs.
}
\description{
Given the list of MOMs, it create the table.
}
