% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTabularReports.R
\name{multiPathwayReport}
\alias{multiPathwayReport}
\title{Summarize pathways' info from a list of MultiOmicsPathway (MOP)}
\usage{
multiPathwayReport(multiPathwayList, priority_to = NULL)
}
\arguments{
\item{multiPathwayList}{MultiOmicsPathway pathway object list}

\item{priority_to}{a vector with the covariates (omic name) that should go first}
}
\value{
a data.frame with overal pvalue of the coxph, followed by covariates zs.
}
\description{
Given the list of MOPs, it create the table.
}
