% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.R
\docType{class}
\name{Omics-class}
\alias{Omics-class}
\alias{Omics}
\title{This class is the storage for the different omic datasets that we need to analyze.}
\description{
This class is the storage for the different omic datasets that we need to analyze.
}
\section{Slots}{

\describe{
\item{\code{data}}{a MultiAssayExperiment containing all the data.}

\item{\code{modelInfo}}{a list with length equal to length(data) that are modelInfo to process each dataset.}

\item{\code{specificArgs}}{a list with length equal to length(data) to set additional parameters specific of the modelInfo.}
}}

