% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{Omics-class}
\alias{Omics-class}
\alias{Omics}
\title{This class is the storage for the different omic datasets that we need to analyze.}
\usage{
Omics(
  experiments = ExperimentList(),
  colData = S4Vectors::DataFrame(),
  sampleMap = S4Vectors::DataFrame(assay = factor(), primary = character(), colname =
    character()),
  metadata = list(),
  drops = list(),
  modelInfo = character(),
  specificArgs = list(),
  pathResult = list()
)
}
\description{
This class is the storage for the different omic datasets that we need to analyze.
}
\section{Slots}{

\describe{
\item{\code{modelInfo}}{a list with length equal to length(data) that are modelInfo to process each dataset.}

\item{\code{specificArgs}}{a list with length equal to length(data) to set additional parameters specific of the modelInfo.}
}}

