% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAggregator-class.R
\docType{class}
\name{Omics-class}
\alias{Omics-class}
\alias{Omics}
\title{Multi Omics Storage that extends \code{"list"} class.}
\usage{
Omics(...)
}
\arguments{
\item{...}{insert the slots. see \code{Slots}}
}
\description{
This class is the storage for the different omic datasets that we need to analyze.
}
\section{Slots}{

\describe{
\item{\code{data}}{list of datasets.}

\item{\code{methods}}{a character vector with length equal to length(data) that are methods to process each dataset.}

\item{\code{specificArgs}}{a list with length equal to length(data) to set additional parameters specific of the methods.}
}}

