% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyGraphModification.R
\name{plotGraphiteInCy}
\alias{plotGraphiteInCy}
\title{Plot graph in Cytoscape}
\usage{
plotGraphiteInCy(graph, nattributes, main = "network")
}
\arguments{
\item{graph}{a graph}

\item{nattributes}{a dataframe with the attribute. Colnames equal to type}

\item{main}{the window title for Cytoscape}
}
\value{
invisble list
\item{gNEL}{the graphNEL object}
\item{cy}{the cytoscape siud}
}
\description{
Given the pathway and nodesAttributes create a graph in Cytoscape
}
