% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotModuleHeat}
\alias{plotModuleHeat}
\title{Plot heatmaps of the module by omics}
\usage{
plotModuleHeat(pathway, moduleNumber, sortBy = NULL, fileName = NULL,
  paletteNames = NULL, h = 9, w = 7)
}
\arguments{
\item{pathway}{MultiOmicsModule pathway object}

\item{moduleNumber}{a module number}

\item{sortBy}{a covariate to sort by}

\item{fileName}{optional filenames to save the plot}

\item{paletteNames}{three palettes}

\item{h}{the height of the plot}

\item{w}{the width of the plot}
}
\description{
Given the pathway and the module, it creates the heatmaps of the mostly involved genes for each omic.
}
