% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingMO.R
\name{plotModuleInGraph}
\alias{plotModuleInGraph}
\title{Plot a Directed Graph of the MultiOmicsModules Object}
\usage{
plotModuleInGraph(
  modulesobj,
  multiOmic,
  reactObj,
  moduleNumber,
  orgDbi = "org.Hs.eg.db",
  paletteNames = NULL,
  legendLabels = NULL,
  fileName = NULL,
  discr_prop_pca = 0.15,
  discr_prop_events = 0.05
)
}
\arguments{
\item{modulesobj}{a \code{MultiOmicsModule} class object}

\item{multiOmic}{an \code{Omics} object}

\item{reactObj}{a \code{PathwayList} from \code{graphite} package that contains the
reactome patways to be used}

\item{moduleNumber}{a module number}

\item{orgDbi}{if needed, indicates an organism Dbi to translate the vectors}

\item{paletteNames}{named vector of MOSpalettes, names replace makeLegend
arguments}

\item{legendLabels}{set up your favourite names for the omics}

\item{fileName}{optional filenames to save the plot}

\item{discr_prop_pca}{the minimal proportion to compute the PCA classes}

\item{discr_prop_events}{the minimal proportion to compute the event classes}
}
\value{
a MOSClip plot in form of a list class object
}
\description{
From a \code{MultiOmicsModules} object, it plots the position of a given module
in the pathway. The omics are also represented in the graph.
}
