% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotModuleInGraph}
\alias{plotModuleInGraph}
\title{Plot KM of the module by omics}
\usage{
plotModuleInGraph(pathway, moduleNumber, makeLegend = NULL, fileName = NULL)
}
\arguments{
\item{pathway}{MultiOmicsModule pathway object}

\item{moduleNumber}{a module number}

\item{makeLegend}{set up your favourite names for the omics}

\item{fileName}{optional filenames to save the plot}
}
\description{
Given the pathway, it creates the Kaplan-meier curves following the formula.
}
