% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotModuleInGraph}
\alias{plotModuleInGraph}
\title{Plot graph of the module by omics}
\usage{
plotModuleInGraph(pathway, moduleNumber, orgDbi = "org.Hs.eg.db",
  paletteNames = NULL, legendLabels = NULL, fileName = NULL,
  discr_prop_pca = 0.15, discr_prop_events = 0.05)
}
\arguments{
\item{pathway}{MultiOmicsModule pathway object}

\item{moduleNumber}{a module number}

\item{orgDbi}{if needed, a organism Dbi to translate vectors}

\item{paletteNames}{named vector of MOpalettes, names replace makeLegend arguments}

\item{legendLabels}{set up your favourite names for the omics}

\item{fileName}{optional filenames to save the plot}

\item{discr_prop_pca}{the minimal proportion to compute the pca classes}

\item{discr_prop_events}{the minimal proportion to compute the event classes}
}
\description{
Given the pathway, it creates the Kaplan-meier curves following the formula.
}
