% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingMO.R
\name{plotModuleKM}
\alias{plotModuleKM}
\title{Plot Kaplan-Meier survival curves of a specific module}
\usage{
plotModuleKM(
  MOM,
  multiOmic,
  moduleNumber,
  formula = "Surv(days, status) ~ PC1",
  fileName = NULL,
  paletteNames = NULL,
  h = 9,
  w = 7,
  risk.table = TRUE,
  pval = TRUE,
  size = 1,
  inYears = FALSE,
  discr_prop_pca = 0.15,
  discr_prop_events = 0.05,
  additional_discrete = NULL,
<<<<<<< HEAD
  additional_continuous = NULL,
  ...
=======
  additional_continuous = NULL
>>>>>>> origin/anna_t
)
}
\arguments{
\item{MOM}{a \code{MultiOmicsModule} class object}

\item{multiOmic}{an \code{Omics} class object}

\item{moduleNumber}{numeric value. The module number of interest}

\item{formula}{a formula for the survival analysis. It should be written as
"Surv(days, status) ~ omic". To plot more than one omic, write them
separated by a "+" character after the separator (~)}

\item{fileName}{optional filenames to save the plot}

\item{paletteNames}{a palette name to be used}

\item{h}{the height of the plot}

\item{w}{the width of the plot}

\item{risk.table}{logical value. If true, shows the risk.table}

\item{pval}{logical value. If true, shows the pvalue}

\item{size}{line width of the KM curves}

\item{inYears}{set time in years}

\item{discr_prop_pca}{the minimal proportion to compute the pca classes}

\item{discr_prop_events}{the minimal proportion to compute the event classes}

\item{additional_discrete}{names of the additional discrete variables to
include}

<<<<<<< HEAD
\item{...}{additional arguments passed to \code{guessInvolvement} and \code{get} function}

\item{additional_continous}{names of the additional continous variables to include}
=======
\item{additional_continuous}{names of the additional continous variables to
include}
}
\value{
a ggsurvplot class object
>>>>>>> origin/anna_t
}
\description{
Given a \code{MultiOmicsModule} class object and a specific module number, it
plots Kaplan-Meier curves, in which the strata corresponds to the omics
}
