% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotModuleKM}
\alias{plotModuleKM}
\title{Plot KM of the module by omics}
\usage{
plotModuleKM(pathway, moduleNumber, formula = "Surv(days, status) ~ PC1",
  fileName = NULL, paletteNames = NULL, h = 9, w = 7,
  risk.table = TRUE, pval = TRUE, size = 1, inYears = FALSE)
}
\arguments{
\item{pathway}{MultiOmicsModule pathway object}

\item{moduleNumber}{a module number}

\item{formula}{a formula to compute the plot}

\item{fileName}{optional filenames to save the plot}

\item{paletteNames}{three palettes}

\item{h}{the height of the plot}

\item{w}{the width of the plot}

\item{risk.table}{logical to show risk.table}

\item{pval}{logical to show pvalue}

\item{size}{line width of the KM curves}

\item{inYears}{set time in years}
}
\description{
Given the pathway, it creates the Kaplan-meier curves following the formula.
}
