% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingMO.R
\name{plotModuleReport}
\alias{plotModuleReport}
\title{Plot a table of a MultiOmicsModule (MOM) object}
\usage{
plotModuleReport(
  modulesObj,
  MOcolors = NULL,
  priority_to = NULL,
  fontsize = 12,
  ...
)
}
\arguments{
\item{modulesObj}{MultiOmicsModules class object}

\item{MOcolors}{character vector with the omic colors.
The colors should be among the colors in \code{showMOSpalette()}}

\item{priority_to}{a vector with the covariates (omic name) that should go
first}

\item{fontsize}{Size of the font to be used in the plot}

\item{...}{additional arguments to be passed to \code{pheatmap}}
}
\value{
a Heatmap list object from ComplexHeatmap package
}
\description{
Given a MOM object, it plots a summary of its results in a tabular fashion
}
