% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotModuleReport}
\alias{plotModuleReport}
\title{Summarize and plot pathways' info from a MultiOmicsModule (MOM) object}
\usage{
plotModuleReport(pathwayObj)
}
\arguments{
\item{pathwayObj}{MultiOmicsModule of pathway object}
}
\description{
Given a MOM, it plots the table.
}
