% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingMO.R
\name{plotMultiPathwayReport}
\alias{plotMultiPathwayReport}
\title{Summarize and plot pathways' info from a list of MultiOmicsPathway (MOP)}
\usage{
plotMultiPathwayReport(
  multiPathwayList,
  top = 25,
  MOcolors = NULL,
  priority_to = NULL,
  fontsize = 6,
  ...
)
}
\arguments{
\item{multiPathwayList}{MultiOmicsPathway list pathway object}

\item{top}{use top number of pathways}

\item{MOcolors}{character vector with the omic colors.
The colors should be among the colors in \code{showMOSpalette()}}

\item{priority_to}{a vector with the covariates (omic name) that should go
first}

\item{fontsize}{the font size to be used}

\item{...}{additional arguments to be passed to \code{pheatmap}}
}
\description{
Given the list of MOPs, it plots the table.
}
