% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingMO.R
\name{plotPathwayHeat}
\alias{plotPathwayHeat}
\title{Plot heatmaps of the pathway by omics}
\usage{
plotPathwayHeat(
  pathway,
  sortBy = NULL,
  paletteNames = NULL,
  additionalAnnotations = NULL,
  additionalPaletteNames = NULL,
  discr_prop_pca = 0.15,
  discr_prop_events = 0.05,
  withSampleNames = TRUE,
  nrowsHeatmaps = 3,
  orgDbi = "org.Hs.eg.db",
  ...
)
}
\arguments{
\item{pathway}{MultiOmicsPathway pathway object}

\item{sortBy}{a covariate to sort by}

\item{paletteNames}{three palettes}

\item{additionalAnnotations}{optional additional sample annotations}

\item{additionalPaletteNames}{optional additional colors for annotations}

\item{discr_prop_pca}{the minimal proportion to compute the pca classes}

\item{discr_prop_events}{the minimal proportion to compute the event classes}

\item{withSampleNames}{create also the samples names}

\item{nrowsHeatmaps}{magnification respect to annotation of sample (annotations take 1 row)}

\item{...}{additional arguments passed to \code{guessInvolvementPathway} function}
}
\description{
Given the pathway, it creates the heatmaps of the mostly involved genes for each omic.
}
