% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotPathwayHeat}
\alias{plotPathwayHeat}
\title{Plot heatmaps of the pathway by omics}
\usage{
plotPathwayHeat(pathway, sortBy = NULL, fileName = NULL,
  paletteNames = c("r_RdYlBu", "BuGn", "Blues"), h = 9, w = 7)
}
\arguments{
\item{pathway}{MultiOmicsPathway pathway object}

\item{sortBy}{a covariate to sort by}

\item{fileName}{optional filenames to save the plot}

\item{paletteNames}{three palettes}

\item{h}{the height of the plot}

\item{w}{the width of the plot}
}
\description{
Given the pathway, it creates the heatmaps of the mostly involved genes for each omic.
}
