% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingMO.R
\name{plotPathwayKM}
\alias{plotPathwayKM}
\title{Plot KM of the pathway by omics}
\usage{
plotPathwayKM(
  pathway,
  multiOmic,
  formula = "Surv(days, status) ~ PC1",
  fileName = NULL,
  paletteNames = NULL,
  h = 9,
  w = 7,
  risk.table = TRUE,
  pval = TRUE,
  size = 1,
  inYears = FALSE,
  discr_prop_pca = 0.15,
  discr_prop_events = 0.05,
  additional_discrete = NULL,
<<<<<<< HEAD
  additional_continous = NULL,
  ...
=======
  additional_continuous = NULL
>>>>>>> origin/anna_t
)
}
\arguments{
\item{pathway}{\code{MultiOmicsModule} class object}

\item{multiOmic}{an \code{Omics} class object}

\item{formula}{a formula to compute the plot}

\item{fileName}{optional filenames to save the plot}

\item{paletteNames}{three palettes}

\item{h}{the height of the plot}

\item{w}{the width of the plot}

\item{risk.table}{logical to show risk.table}

\item{pval}{logical to show pvalue}

\item{size}{line width of the KM curves}

\item{inYears}{set time in years}

\item{discr_prop_pca}{the minimal proportion to compute the pca classes}

\item{discr_prop_events}{the minimal proportion to compute the event classes}

\item{additional_discrete}{names of the additional discrete variables to
include}

<<<<<<< HEAD
\item{additional_continous}{names of the additional continous variables to include}

\item{...}{additional arguments passed to \code{guessInvolvementPathway} and \code{get} function}
=======
\item{additional_continuous}{names of the additional continous variables to
include}
>>>>>>> origin/anna_t
}
\description{
Given the pathway, it creates the Kaplan-meier curves following the formula.
}
