% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMO.R
\name{plotPathwayKM}
\alias{plotPathwayKM}
\title{Plot KM of the pathway by omics}
\usage{
plotPathwayKM(pathway, formula = "Surv(days, status) ~ PC1",
  fileName = NULL, paletteName = c("r_RdYlBu", "BuGn", "Blues"), h = 9,
  w = 7)
}
\arguments{
\item{pathway}{MultiOmicsModule pathway object}

\item{formula}{a formula to compute the plot}

\item{fileName}{optional filenames to save the plot}

\item{h}{the height of the plot}

\item{w}{the width of the plot}

\item{moduleNumber}{a module number}

\item{paletteNames}{three palettes}
}
\description{
Given the pathway, it creates the Kaplan-meier curves following the formula.
}
