% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis-functions.R
\name{plotSurvivalBarcodes}
\alias{plotSurvivalBarcodes}
\title{Plot the patients barcodes}
\usage{
plotSurvivalBarcodes(coxDiscrete, worstProfile, filename = NA)
}
\arguments{
\item{coxDiscrete}{the discrete table genes x samples}

\item{worstProfile}{a data.frame as produced by extractBadPrognosisProfile}

\item{filename}{if NA plot in session otherwise plot in filename}
}
\description{
Given the patients' classes of the genes and the worst prognosis, the function produces a plot of the barcodes.
}
