% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactSmall}
\alias{reactSmall}
\title{PathwayList of pathways from Reactome}
\format{
\subsection{\code{reactSmall}}{

A PathwayList with Reactome pathways for hsapiens
\describe{
\item{entries}{Three Reactome pathways with their nodes}
}
}
}
\usage{
data('reactSmall')
}
\description{
A PathwayList with three pathways necessary for the analysis:
'Activation of Matrix Metalloproteinases', 'FGFR1 mutant receptor
activation', and 'VEGFA-VEGFR2 Pathway'. Pathways were downloaded using
\code{graphite} package and the names of the nodes were converted into
Entrez IDs.
}
\keyword{datasets}
