% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingFunctions.R
\name{checkOrder}
\alias{checkOrder}
\alias{resolveAndOrder}
\alias{mergeCol}
\alias{filterExpr}
\alias{filterMultiOmicsForSamples}
\alias{preparePerms}
\title{Check if all the list object have the same order of pathway module}
\usage{
checkOrder(li)

resolveAndOrder(li)

mergeCol(li, col = "PC1", resolve = FALSE)

filterExpr(exp, samples)

filterMultiOmicsForSamples(MO, samples)

preparePerms(fullMultiOmics, nperm = 100, nPatients = 3)
}
\arguments{
\item{li}{a list of summaries}

\item{col}{the column to merge}

\item{resolve}{weather to resolve the issues}

\item{exp}{a matrix}

\item{samples}{the vector of samples to select}

\item{MO}{a multiOmic object}

\item{fullMultiOmics}{a multiOmic object}

\item{nperm}{number of permutations}

\item{nPatients}{number of patients to remove for resampling}
}
\value{
a matrix

a filtered matrix

a filtered MultiOmics objects

list of sampled patients for each permutation
}
\description{
For internal use only

For internal use only

For internal use only

For internal use only

For internal use only

Prepare subset of patients for permutations
}
