% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{showModule}
\alias{showModule}
\title{A generic function showing pathway's module info}
\usage{
showModule(object)
}
\arguments{
\item{object}{an object of class \code{MultiOmicsModules}}
}
\description{
A generic function showing pathway's module info
}
\examples{
data(multiOmics)
data(reactSmall)

genesToUse <- row.names(multiOmics[[1]])

MOM_survival <- multiOmicsSurvivalModuleTest(multiOmics, reactSmall[[1]],
  survFormula="Surv(days, status) ~", autoCompleteFormula = TRUE,
  useTheseGenes = genesToUse)
  
showModule(MOM_survival)

}
