% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{showMOSpalette}
\alias{showMOSpalette}
\title{Shows the MOSClip palette.}
\usage{
showMOSpalette()
}
\value{
NULL. No value is returned
}
\description{
This function shows the MOSClip palette.
Each omic should be coupled to a color panel, this match will be preserved
in plots.
}
\examples{
showMOSpalette()

}
