% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{showOmics}
\alias{showOmics}
\title{A generic functions showing parameter associated with each omics}
\usage{
showOmics(object)
}
\arguments{
\item{object}{an object of class \code{Omics}}
}
\description{
A generic functions showing parameter associated with each omics
}
