% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{showPathway}
\alias{showPathway}
\title{A generic function showing pathway info}
\usage{
showPathway(object)
}
\arguments{
\item{object}{an object of class \code{MultiOmicsPathway}}
}
\description{
A generic function showing pathway info
}
\examples{
data(multiOmics)
data(reactSmall)

genesToUse <- row.names(multiOmics[[1]])

MOP_survival <- multiOmicsSurvivalPathwayTest(multiOmics, reactSmall[[1]],
  survFormula="Surv(days, status) ~", autoCompleteFormula = TRUE,
  useTheseGenes = genesToUse)
  
showPathway(MOP_survival)

}
