% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{showPathway}
\alias{showPathway}
\title{A generic function showing pathway info}
\usage{
showPathway(object)
}
\arguments{
\item{object}{an object of class MultiOmicsPathway}
}
\description{
A generic function showing pathway info
}
