% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertest.R
\name{stripModulesFromPathways}
\alias{stripModulesFromPathways}
\title{Remove module number From Pathway Name}
\usage{
stripModulesFromPathways(pathways)
}
\arguments{
\item{pathways}{vector of pathway names}
}
\description{
Remove module number From Pathway Name
}
