% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicMethods.R
\name{summarizeInCluster}
\alias{summarizeInCluster}
\title{Summarize Using Cluster Analysis}
\usage{
summarizeInCluster(data, features, name = "clu", dictionary = NULL,
  max_cluster_number = 3, cliques = NULL)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{dictionary}{translate features (genes) into sets (row.names of the data)}

\item{max_cluster_number}{the maximum number of cluster to evaluate}

\item{cliques}{the features organized in cliques. Only use for topology}
}
\description{
Given a matrix it summarize in classes
}
\details{
The user can define a maximum of classes. The function
guess the optimal number of clusters using NbClust methods.
}
