% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicsMethod.R
\name{summarizeToBinaryDirectionalEvents}
\alias{summarizeToBinaryDirectionalEvents}
\title{Summarize To Binary Directional Events}
\usage{
summarizeToBinaryDirectionalEvents(
  data,
  features,
  name = "dirBin",
  binaryClassMin = 10,
  eventThr = 2,
  cliques = NULL
)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{binaryClassMin}{the minimum number of event to include the covariate}

\item{eventThr}{the absolute value to threshold an event}

\item{cliques}{the features organized in cliques. Only use for topology}
}
\value{
a list with summary of the omic:
\item{x}{summary of the omic for each sample}
\item{usedGenes}{genes list of genes used to calculate the summary}
\item{namesCov}{names of the covariates}
\item{method}{method used for the analysis}
\item{omicName}{name of the omic}
\item{evenThr}{threshold fot event counting}
}
\description{
Given a matrix it summarize the positive and negative to 0 or 1
in two vectors
}
