% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicsMethod.R
\name{summarizeToNumberOfEvents}
\alias{summarizeToNumberOfEvents}
\title{Summarize To Number of Binary Events}
\usage{
summarizeToNumberOfEvents(
  data,
  features,
  name = "event",
  min_prop = 0.1,
  cliques = NULL
)
}
\arguments{
\item{data}{a data matrix}

\item{features}{a vector with the features to analyze}

\item{name}{prefix of the covariates}

\item{min_prop}{minimal proportion in classes}

\item{cliques}{the features organized in cliques. Only use for topology}
}
\value{
a list with summary of the omic:
\item{x}{summary of the omic for each sample}
\item{usedGenes}{genes list of genes used to calculate the summary}
\item{namesCov}{names of the covariates}
\item{method}{method used for the analysis}
\item{omicName}{name of the omic}
\item{evenThr}{threshold fot event counting}
\item{min_prop}{minimum proportion of samples to exclude to check the
variability of values}
}
\description{
Given a matrix it summarize to a 0 or 1
}
